requirements.txt
numpy>=1.19.1,<1.20
pyglet>=1.5.21,<1.6
pywavefront>=1.3.3,<1.4
moderngl>=5.6.4,<5.7
moderngl-window>=2.4.0,<2.5
pyrr>=0.10.3,0.11
pillow>=8.4.0,<8.5